refactor: rename jest presets (#4609)
* refactor: rename jest presets

Follow up of #4597.
Renamed `preset` folder to `presets`.
Added a default preset and it is accessible in the short form of
`@vue/cli-plugin-unit-jest`.

This change is inspired by ts-jest presets:
https://kulshekhar.github.io/ts-jest/user/config/#jest-preset

* fix: update import path

* fix: fix typo in file path
