+++ packages/@vue/cli-plugin-unit-jest/jest-preset.js
@@ -1,0 +1,1 @@
+module.exports = require('./presets/default/jest-preset')
