+++ packages/@vue/cli-plugin-unit-jest/presets/default/jest-preset.js
