+++ packages/@vue/cli-plugin-unit-jest/presets/no-babel/esmoduleTransformer.js
